* main: (556 commits)
  Update README.md
  Update IconDatabase.cs
  Update Repo URL
  Load updated translations from Tolgee Co-authored-by: Martí Climent from the multiverse <121457539+martinet101@users.noreply.github.com>
  Fix a small issue that prevented passing Microsoft Store validation
  Update README.md
  Update README.md
  Update icons and screenshots from the excel file
  Load updated translations from Tolgee Co-authored-by: Martí Climent from the multiverse <121457539+martinet101@users.noreply.github.com>
  Update icons and screenshots from the excel file
  Load updated translations from Tolgee Co-authored-by: Martí Climent from the multiverse <121457539+martinet101@users.noreply.github.com>
  Update cli-arguments.md
  Revert "Disable Icons on package lists (they are not ready yet)"
  Update icons and screenshots from the excel file
  Load updated translations from Tolgee Co-authored-by: Martí Climent from the multiverse <121457539+martinet101@users.noreply.github.com>
  Update version number and translations
  fix scoop packages installed from url breaking on UniGetUI (fix #2853)
  fix scoop packages installed from url breaking on UniGetUI (fix #2853)
  Fix some WinGet packages not being properly uninstalled
  Fix some german words not displaying properly (fix #2843)
  ...

+++ src/UniGetUI.Core.Settings/SettingsEngine.cs
@@ -1,33 +1,49 @@
-﻿using UniGetUI.Core.Data;
+using System.Collections.Concurrent;
+using UniGetUI.Core.Data;
 using UniGetUI.Core.Logging;
 
 namespace UniGetUI.Core.SettingsEngine
 {
     public static class Settings
     {
+        private static ConcurrentDictionary<string, bool> booleanSettings = new();
+        private static ConcurrentDictionary<string, string> valueSettings = new();
+
         public static bool Get(string setting, bool invert = false)
         {
-            return File.Exists(Path.Join(CoreData.UniGetUIDataDirectory, setting)) ^ invert;
+            if (booleanSettings.TryGetValue(setting, out bool result))
+            {   // If the setting was cached
+                return result ^ invert;
+            }
+
+            // Otherwhise, load the value from disk and cache that setting
+            result = File.Exists(Path.Join(CoreData.UniGetUIDataDirectory, setting));
+            booleanSettings[setting] = result;
+            return result ^ invert;
         }
 
         public static void Set(string setting, bool value)
         {
             try
             {
-                if (value)
+                // Cache that setting's new value
+                booleanSettings[setting] = value;
+
+                // Update changes on disk if applicable
+                if (value && !File.Exists(Path.Join(CoreData.UniGetUIDataDirectory, setting)))
                 {
-                    if (!File.Exists(Path.Join(CoreData.UniGetUIDataDirectory, setting)))
-                    {
-                        File.WriteAllText(Path.Join(CoreData.UniGetUIDataDirectory, setting), "");
-                    }
+                    File.WriteAllText(Path.Join(CoreData.UniGetUIDataDirectory, setting), "");
                 }
-                else
+                else if (!value)
                 {
+                    valueSettings[setting] = "";
+
                     if (File.Exists(Path.Join(CoreData.UniGetUIDataDirectory, setting)))
                     {
                         File.Delete(Path.Join(CoreData.UniGetUIDataDirectory, setting));
                     }
                 }
+
             }
             catch (Exception e)
             {
@@ -38,25 +54,38 @@ namespace UniGetUI.Core.SettingsEngine
 
         public static string GetValue(string setting)
         {
-            if (!File.Exists(Path.Join(CoreData.UniGetUIDataDirectory, setting)))
-            {
-                return "";
+            if (valueSettings.TryGetValue(setting, out string? value))
+            {   // If the setting was cached
+                return value;
             }
 
-            return File.ReadAllText(Path.Join(CoreData.UniGetUIDataDirectory, setting));
+            // Otherwhise, load the setting from disk and cache that setting
+            value = "";
+            if (File.Exists(Path.Join(CoreData.UniGetUIDataDirectory, setting)))
+            {
+                value = File.ReadAllText(Path.Join(CoreData.UniGetUIDataDirectory, setting));
+            }
+
+            valueSettings[setting] = value;
+            return value;
+
         }
 
         public static void SetValue(string setting, string value)
         {
             try
             {
-                if (value == "")
+                if (value == String.Empty)
                 {
                     Set(setting, false);
+                    booleanSettings[setting] = false;
+                    valueSettings[setting] = "";
                 }
                 else
                 {
                     File.WriteAllText(Path.Join(CoreData.UniGetUIDataDirectory, setting), value);
+                    booleanSettings[setting] = true;
+                    valueSettings[setting] = value;
                 }
             }
             catch (Exception e)
@@ -64,6 +93,36 @@ namespace UniGetUI.Core.SettingsEngine
                 Logger.Error($"CANNOT SET SETTING VALUE FOR setting={setting} enabled={value}");
                 Logger.Error(e);
             }
+        }
+
+        /*
+         *
+         *
+         */
+
+        public static bool AreNotificationsDisabled()
+        {
+            return Get("DisableSystemTray") || Get("DisableNotifications");
+        }
+
+        public static bool AreUpdatesNotificationsDisabled()
+        {
+            return AreNotificationsDisabled() || Get("DisableUpdatesNotifications");
+        }
+
+        public static bool AreErrorNotificationsDisabled()
+        {
+            return AreNotificationsDisabled() || Get("DisableErrorNotifications");
+        }
+
+        public static bool AreSuccessNotificationsDisabled()
+        {
+            return AreNotificationsDisabled() || Get("DisableSuccessNotifications");
+        }
+
+        public static bool AreProgressNotificationsDisabled()
+        {
+            return AreNotificationsDisabled() || Get("DisableProgressNotifications");
         }
     }
 }
